This patch contains two completely unrelated changes. The first change is to the lcm and gcd functions so that they accept both lists and tuples (previously tuples were not accepted).
The more important change is to the random_prime() function. The previous
implementation choose a number at random and then returned the first prime
below it, which does not choose primes uniformly. This implementation
is fixed so that it now does. The time penalty is only (expected)
logarithmic in n.