Fixed rank bug that Nick Alexander found.
When we switched to linbox the linbox rank algorithm got called and
didn't raise an error.  I've changed SAGE to now only call that linbox
routine when the base ring is a field.  Otherwise it calls a generic
routine, which, in this case, properly raises a NotImplementedError.
See attached patch.