workaround for Singular-3-0-4-2 changes
Singular implements finite extension fields as univariate polynomial rings over
the base field. We used to setup these wrong in particular in a way such that
the elements had components, which they are not supposed to have. This patches
fixes that (rather hackish) until the Singular team advised us on a proper way to do this.

This fixes the remaining GCD problems with Singular-3-0-4-2 in #2772.