For #5929: fix hang in maxima/ecl due to assumption questions
When maxima asks a question, current code sends CTRL-C
twice to break, then raises an exception. This used to work with
clisp, but for ecl it actually hangs. A test case is given by

{{{
var("Ax,Bx,By")
t = -Ax*sin(sqrt(Ax^2)/2)/(sqrt(Ax^2)*sqrt(By^2 + Bx^2))
limit(t, Ax=0)
}}}

It turns out that it is possible to "break" from the question by just
sending ";" by itself to maxima. The current patch changes to this
method of escaping. It works with clisp or ecl.