#9896 Fix/add dependencies of sage/libs/{mwrank, cremona/*} (didn't get rebuilt after ECLIB update/rebuild).
I've not (yet) removed the IMHO unnecessary PARI library from three of these
modules' library lists. Having it there (also) caused trouble in upgrades to
Sage versions with a new PARI library, because the *old* one was still recorded
as a prerequisite in the extension modules' shared / dynamic libraries, s.t.
either the old one wasn't found (ImportError) or the wrong (old) one got loaded
(leading to segfaults etc.).

See also #9914; it is claimed there that Cygwin needs this. (If so, we should
use uname_specific() to only include the library on that platform.)

If we keep it there, we should also make these modules depend on e.g.
SAGE_LOCAL/include/pari/pari.h, though they *now* indirectly depend on PARI
through the added dependencies on ECLIB (which in turn depends on PARI).